dedx float, 2 IK steps

+++ wasm/include/GLTF.h
@@ -182,10 +182,10 @@ class GLTF : public OptimizationProblem{
         static glm::vec3 applyRotation(const glm::vec3 x, const glm::quat rot);
 
         // Computes the gradient of a rotation's quaternion with respect to an error given gradient of x output to that error
-        static glm::vec4 dedq(const glm::vec3 x, const glm::quat rot, const glm::dvec3 dedx);
+        static glm::vec4 dedq(const glm::vec3 x, const glm::quat rot, const glm::vec3 dedx);
 
         // Computes the gradient of a rotation's input with respect to an error given gradient of x output to that error
-        static glm::dvec3 dedx(const glm::vec3 x, const glm::quat rot, const glm::dvec3 dedx);
+        static glm::vec3 dedx(const glm::vec3 x, const glm::quat rot, const glm::vec3 dedx);
 
         // hashes a vertex to allow duplicates to be detected
         int hashVertex(glm::vec3 v);
